backend/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
cloudinary==1.37.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-import-export==3.3.7
django-rest-framework==0.1.0
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.1.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
et-xmlfile==1.1.0
Faker==23.1.0
filelock==3.13.1
fsspec==2024.2.0
idna==3.6
inflection==0.5.1
Jinja2==3.1.3
joblib==1.3.2
MarkupPy==1.14
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.2.0
Pillow==10.0.0
psycopg2==2.9.9
PyJWT==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
scikit-learn==1.4.0
scipy==1.12.0
setuptools==70.0.0
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tablib==3.5.0
threadpoolctl==3.2.0
toposort==1.10
torch==2.2.1
torchvision==0.17.1
typing_extensions==4.10.0
tzdata==2023.3
urllib3==2.1.0
xlrd==2.0.1
xlwt==1.3.0
